add generator.py file with class and wrapper
